Minor changes.
This should be the end of "phase 1" of the KAI library. Where I simply just copy and paste the MVPs' base code and rebrand it, with some minor fixes and changes here and there. Next begins the process of moving base vehicle code to KAI_Actor itself, stuff like virtuals such as CanAttack and ShouldAttack.