Partially reworked KAI_AvoidObstacles().
Rewrote KAI_AvoidObstacles to get ALL valid positions around the caller, and pick the furthest one. Because it was more effective than trying to make the old "Find first unblocked check to the left or right" check to work, since there's I guess no way to check if it should check clockwise or counterclockwise.