Added basic obstacle climbing.
KMT_3D flyers will now climb up and down obstacles they encounter. Like climb up raised floors and down lowered ceilings. For actors, they'll check if either the bottom or top of the blocker are free, if both are, it'll go whichever way is nearest from the middle height of the caller. If neither are free, it treats it as a full block.

Next I'll also add handling for other finite height obstacles like 3D floors and 3D midtexes, using the same logic as for actors.